add typedef Py::TopoShape for convenience, move shape2pyshape to own file
