+++ src/Mod/Part/App/PartPyCXX.h
@@ -1,0 +1,38 @@
+/***************************************************************************
+ *   Copyright (c) 2017 Werner Mayer <wmayer[at]users.sourceforge.net>     *
+ *                                                                         *
+ *   This file is part of the FreeCAD CAx development system.              *
+ *                                                                         *
+ *   This library is free software; you can redistribute it and/or         *
+ *   modify it under the terms of the GNU Library General Public           *
+ *   License as published by the Free Software Foundation; either          *
+ *   version 2 of the License, or (at your option) any later version.      *
+ *                                                                         *
+ *   This library  is distributed in the hope that it will be useful,      *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU Library General Public License for more details.                  *
+ *                                                                         *
+ *   You should have received a copy of the GNU Library General Public     *
+ *   License along with this library; see the file COPYING.LIB. If not,    *
+ *   write to the Free Software Foundation, Inc., 59 Temple Place,         *
+ *   Suite 330, Boston, MA  02111-1307, USA                                *
+ *                                                                         *
+ ***************************************************************************/
+
+#ifndef PART_PYCXX_H
+#define PART_PYCXX_H
+
+#include <CXX/Extensions.hxx>
+#include <Mod/Part/App/TopoShapePy.h>
+
+namespace Py {
+    typedef ExtensionObject<Part::TopoShapePy> TopoShape;
+    template<>
+    bool TopoShape::accepts (PyObject *pyob) const
+    {
+        return (pyob && PyObject_TypeCheck(pyob, &(Part::TopoShapePy::Type)));
+    }
+}
+
+#endif //PART_PYCXX_H
